refactor: set main page to home page
